src/main/java/com/mindex/challenge/data/Compensation.java
package com.mindex.challenge.data;

import java.util.Date;

public class Compensation {
    private String employeeId;
    private float salary;
    private Date effectiveDate;

    public Compensation() {
    }

    public Compensation(String employeeId, float salary, Date effectiveDate) {
        this.employeeId = employeeId;
        this.salary = salary;
        this.effectiveDate = effectiveDate;
    }

    public String getEmployeeId() {
        return employeeId;
    }

    public void setEmployeeId(String employeeId) {
        this.employeeId = employeeId;
    }

    public float getSalary() {
        return salary;
    }

    public void setSalary(float salary) {
        this.salary = salary;
    }

    public Date getEffectiveDate() {
        return effectiveDate;
    }

    public void setEffectiveDate(Date effectiveDate) {
        this.effectiveDate = effectiveDate;
    }
}
